XCBhsb kjhfkqjad

+++ src/main/java/com/v1/planner/service/TravelPlanService.java
@@ -3,9 +3,10 @@ package com.v1.planner.service;
 import com.v1.planner.dto.TravelPlanRequest;
 import com.v1.planner.entity.TravelPlan;
 import com.v1.planner.repository.TravelPlanRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -15,12 +16,18 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
 public class TravelPlanService {
-    @Autowired
-    private TravelPlanRepository travelPlanRepository;
+    private final TravelPlanRepository travelPlanRepository;
+
+    public TravelPlanService(TravelPlanRepository travelPlanRepository) {
+        this.travelPlanRepository = travelPlanRepository;
+    }
+
+
 
     @Value("${upload.path}")
     private String uploadDir;
@@ -29,10 +36,13 @@ public class TravelPlanService {
         return travelPlanRepository.findAll();
     }
 
+
     public Optional<TravelPlan> getTravelPlanById(Long id) {
-        return travelPlanRepository.findById(id);
+        return Optional.ofNullable(travelPlanRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("Travel plan not found with id: " + id)));
     }
 
+    @Transactional
     public TravelPlan createTravelPlan(TravelPlanRequest travelPlanRequest, List<MultipartFile> files) throws IOException {
         List<String> filePaths = new ArrayList<>();
 
@@ -62,6 +72,7 @@ public class TravelPlanService {
         return travelPlanRepository.save(travelPlan);
     }
 
+    @Transactional
     public TravelPlan updateTravelPlan(Long id, TravelPlan travelPlan) {
         Optional<TravelPlan> existingTravelPlanOpt = travelPlanRepository.findById(id);
         if (existingTravelPlanOpt.isPresent()) {
@@ -77,10 +88,14 @@ public class TravelPlanService {
         }
     }
 
+
+    @Transactional
     public void deleteTravelPlan(Long id) {
+        if (!travelPlanRepository.existsById(id)) {
+            throw new NoSuchElementException("Cannot delete. Travel plan not found with id: " + id);
+        }
         travelPlanRepository.deleteById(id);
     }
-
     public TravelPlan buyTravelPlan(Long userId, Long travelPlanId) {
         Optional<TravelPlan> travelPlanOpt = travelPlanRepository.findById(travelPlanId);
         if (travelPlanOpt.isPresent()) {
@@ -94,4 +109,6 @@ public class TravelPlanService {
             throw new RuntimeException("Travel plan not found");
         }
     }
+
+
 }
